php/Deployment Test/Deployment Test/requirements.txt
Flask==2.2.2
gunicorn==19.6.0
Werkzeug=2.2.2
PyMySQL=1.0.2
Flask-MySQL=1.5.2
Flask-SQLAlchemy=3.0.2
google-auth=1.33.0
google-cloud-core=1.7.1
google-cloud-storage=1.31.0
